+++ pages/index.js
@@ -17,283 +17,128 @@ import {
   Stack,
   Button,
   TextContainer,
-  Pagination
+  Pagination,
+  Caption,
+  Banner,
+  List,
+  DropZone,
+  FileUpload,
+  Collapsible
 } from "@shopify/polaris";
 import React from "react";
 // import { Link } from "@shopify/polaris/types/components/AppProvider/utilities";
-
+// 123123eqweqwe
 class Index extends React.Component {
-  render() {
+  constructor(state) {
+    super(state);
+    this.state = {
+      searchValue: "",
+      appliedFilters: [
+        {
+          key: "accountStatusFilter",
+          value: "Account enabled"
+        }
+      ]
+    };
+  }
+
+  handleSearchChange = searchValue => {
+    this.setState({ searchValue });
+  };
+
+  handleFiltersChange = appliedFilters => {
+    this.setState({ appliedFilters });
+  };
+
+  renderItem = item => {
+    const { id, url, name, location } = item;
+    const media = <Avatar customer size="medium" name={name} />;
+
     return (
-      <Page title="Phomas">
-        <Layout>
-          <Layout.Section>
-            <Card sectioned>
-              <Card.Section>
-                <Stack>
-                  <Stack.Item fill>
-                    <Heading element="h1">Welcome to,</Heading>
-                    <DisplayText size="extraLarge">Phomas</DisplayText>
-                  </Stack.Item>
-                  <Stack.Item>
-                    <div style={{ height: "100px" }}>
-                      <Pagination
-                        hasPrevious
-                        previousKeys={[74]}
-                        onPrevious={() => {
-                          console.log("Previous");
-                        }}
-                        hasNext
-                        nextKeys={[75]}
-                        onNext={() => {
-                          console.log("Next");
-                        }}
-                      />
-                    </div>
-                  </Stack.Item>
-                </Stack>
-              </Card.Section>
-              <Card.Section>
-                <TextContainer spacing="loose">
-                  <p>
-                    To empower your e-commerce sales and reduce costs, Linx
-                    developed Phomas. A unique solution to manage the process of
-                    E-commerce photography. Including an online retouch and
-                    approval workflow and automated image conversion and
-                    delivery. Get the max out of your web shop and reduce
-                    returns by displaying the most photo realistic imagery.
-                    Boost sales by speeding up the process. Phomas delivers!
-                    Phomas is an integrated and highly configurable Cloud
-                    Solution. Connect your photographers and your photo retouch
-                    company. Deliver to your own web shop or these of
-                    affiliates. Auto crop imagery based on business logic.
-                    Phomas gives you the control you need to benifit most of
-                    your e-commerce channel.
-                  </p>
-                </TextContainer>
-                <Layout>
-                  <video
-                    autoPlay="autoplay"
-                    loop="loop"
-                    controls="controls"
-                    width="500"
-                    height="400"
-                  >
-                    <source
-                      type="video/mp4"
-                      src="https://phomas.com/Media/Default/Video/Phomas-PhotoShootManagement-HD.mp4"
-                    />
-                  </video>
-                </Layout>
-              </Card.Section>
-              <Layout>
-                <Button plain>
-                  Get more detailed information at phomas.com
-                </Button>
-              </Layout>
-            </Card>
-          </Layout.Section>
-        </Layout>
+      <Card>
+        <Card.Header
+          actions={[
+            {
+              content: "-Coming soon-"
+            }
+          ]}
+          title="Multiple Variant Images"
+        />
+        <Collapsible>
+          <Card.Section>
+            <Card />
+          </Card.Section>
+        </Collapsible>
+      </Card>
+    );
+  };
 
-        <Layout>
-          <Layout.Section>
-            <Card sectioned>
-              <Card.Section>
-                <Stack alignment="center">
-                  <Stack.Item fill>
-                    <DisplayText size="extraLarge">Functionalities</DisplayText>
-                  </Stack.Item>
-                  <Stack.Item>
-                    <div style={{ height: "100px" }}>
-                      <Pagination
-                        hasPrevious
-                        previousKeys={[74]}
-                        onPrevious={() => {
-                          console.log("Previous");
-                        }}
-                        hasNext
-                        nextKeys={[75]}
-                        onNext={() => {
-                          console.log("Next");
-                        }}
-                      />
-                    </div>
-                  </Stack.Item>
-                </Stack>
-              </Card.Section>
-              <Card.Section>
-                {/* ------------------------------------ */}
-                <TextContainer spacing="loose">
-                  <Layout section={true}>
-                    <Stack alignment="center">
-                      <Layout.Section>
-                        <Heading element="h1">Main image</Heading>
-                        <br />
-                      </Layout.Section>
-                    </Stack>
+  render() {
+    const resourceName = {
+      singular: "customer",
+      plural: "customers"
+    };
 
-                    <p>
-                      The first image for each product is known as the featured,
-                      or main, image. The featured image is shown on collection
-                      pages, the cart page, the checkout page, and your home
-                      page. Product images are displayed on your online store to
-                      let your customers see what your products look like. The
-                      images are also used on Shopify POS and other sales
-                      channels. The best file type for most product images is
-                      JPEG.
-                    </p>
-                  </Layout>
-                </TextContainer>
-                <br />
-                <hr />
-                <br />
-                {/* ------------------------------------ */}
-                <TextContainer spacing="loose">
-                  <Layout section={true}>
-                    <Stack alignment="center">
-                      <Layout.Section>
-                        <Heading element="h1">H-over</Heading>
-                        <br />
-                      </Layout.Section>
-                    </Stack>
-                    <p>
-                      Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                      Proin eros sapien, lacinia at lacus sit amet, aliquet
-                      malesuada diam. Integer congue nibh non ipsum ornare
-                      iaculis. Nulla ut scelerisque tortor. Nulla pharetra sem
-                      rutrum diam ornare congue. Fusce eget ullamcorper massa.
-                      Mauris porttitor dictum dapibus. Ut faucibus varius
-                      pharetra. Proin sed rutrum odio.
-                    </p>
-                  </Layout>
-                </TextContainer>
-                <br />
-                <hr />
-                <br />
-                {/* -------------------------------- */}
-                <TextContainer spacing="loose">
-                  <Layout section={true}>
-                    <Stack alignment="center">
-                      <Layout.Section>
-                        <Heading element="h1">
-                          Add multiple variant images
-                        </Heading>
-                        <br />
-                      </Layout.Section>
-                    </Stack>
-                    <p>
-                      Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                      Proin eros sapien, lacinia at lacus sit amet, aliquet
-                      malesuada diam. Integer congue nibh non ipsum ornare
-                      iaculis. Nulla ut scelerisque tortor. Nulla pharetra sem
-                      rutrum diam ornare congue. Fusce eget ullamcorper massa.
-                      Mauris porttitor dictum dapibus. Ut faucibus varius
-                      pharetra. Proin sed rutrum odio.
-                    </p>
-                  </Layout>
-                  <br />
-                  <br />
-                </TextContainer>
-                <Layout>
-                  <Button plain>Get the code snippets</Button>
-                </Layout>
-                {/* -------------------------------- */}
-              </Card.Section>
-            </Card>
-          </Layout.Section>
-        </Layout>
+    const items = [
+      {
+        id: 341,
+        url: "customers/341",
+        name: "Mae Jemison",
+        location: "Decatur, USA"
+      },
+      {
+        id: 256,
+        url: "customers/256",
+        name: "Ellen Ochoa",
+        location: "Los Angeles, USA"
+      }
+    ];
 
-        <Layout>
-          <Layout.Section>
-            <Card sectioned>
-              <Card.Section>
-                <Stack alignment="center">
-                  <Stack.Item fill>
-                    <DisplayText size="extraLarge">
-                      Final information
-                    </DisplayText>
-                  </Stack.Item>
-                  <Stack.Item>
-                    <div style={{ height: "100px" }}>
-                      <Pagination
-                        hasPrevious
-                        previousKeys={[74]}
-                        onPrevious={() => {
-                          console.log("Previous");
-                        }}
-                        hasNext
-                        nextKeys={[75]}
-                        onNext={() => {
-                          console.log("Next");
-                        }}
-                      />
-                    </div>
-                  </Stack.Item>
-                </Stack>
-              </Card.Section>
-              <Card.Section>
-                {/* ------------------------------------ */}
-                <TextContainer spacing="loose">
-                  <Layout section={true}>
-                    <Stack alignment="center">
-                      <Layout.Section>
-                        <Heading element="h1">File naming</Heading>
-                        <br />
-                      </Layout.Section>
-                    </Stack>
+    const filters = [
+      {
+        key: "orderCountFilter",
+        label: "Number of orders",
+        operatorText: "is greater than",
+        type: FilterType.TextField
+      },
+      {
+        key: "accountStatusFilter",
+        label: "Account status",
+        operatorText: "is",
+        type: FilterType.Select,
+        options: ["Enabled", "Invited", "Not invited", "Declined"]
+      }
+    ];
 
-                    <p>
-                      Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                      Proin eros sapien, lacinia at lacus sit amet, aliquet
-                      malesuada diam. Integer congue nibh non ipsum ornare
-                      iaculis. Nulla ut scelerisque tortor. Nulla pharetra sem
-                      rutrum diam ornare congue. Fusce eget ullamcorper massa.
-                      Mauris porttitor dictum dapibus. Ut faucibus varius
-                      pharetra. Proin sed rutrum odio.
-                    </p>
-                  </Layout>
-                </TextContainer>
-                <br />
-                <hr />
-                <br />
-                {/* ------------------------------------ */}
-                <TextContainer spacing="loose">
-                  <Layout section={true}>
-                    <Stack alignment="center">
-                      <Layout.Section>
-                        <Heading element="h1">
-                          Information about the trial version
-                        </Heading>
-                        <br />
-                      </Layout.Section>
-                    </Stack>
-                    <p>
-                      Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                      Proin eros sapien, lacinia at lacus sit amet, aliquet
-                      malesuada diam. Integer congue nibh non ipsum ornare
-                      iaculis. Nulla ut scelerisque tortor. Nulla pharetra sem
-                      rutrum diam ornare congue. Fusce eget ullamcorper massa.
-                      Mauris porttitor dictum dapibus. Ut faucibus varius
-                      pharetra. Proin sed rutrum odio.
-                    </p>
-                  </Layout>
-                </TextContainer>
-                <br />
-                <br />
-                {/* -------------------------------- */}
-                <Layout>
-                  <Button size="large" primary={true}>
-                    Upload image from pc
-                  </Button>
-                </Layout>
-                {/* -------------------------------- */}
-              </Card.Section>
-            </Card>
-          </Layout.Section>
-        </Layout>
+    const filterControl = (
+      <ResourceList.FilterControl
+        filters={filters}
+        appliedFilters={this.state.appliedFilters}
+        onFiltersChange={this.handleFiltersChange}
+        searchValue={this.state.searchValue}
+        onSearchChange={this.handleSearchChange}
+        additionalAction={{
+          content: "Save",
+          onAction: () => console.log("New filter saved")
+        }}
+      />
+    );
+
+    return (
+      <Page fullWidthtitle="Ph2Shopify" forceRender>
+        <Stack distribution="trailing">
+          <Button primary>Upload</Button>
+        </Stack>
+        <Card sectioned={true}>
+          <ResourceList
+            resourceName={resourceName}
+            items={items}
+            renderItem={this.renderItem}
+            filterControl={filterControl}
+          />
+        </Card>
       </Page>
     );
   }
 }
-
 export default Index;
